+++ src/cmd/root.go
@@ -23,7 +23,7 @@ import (
 	"github.com/zarf-dev/zarf/src/pkg/cluster"
 	"github.com/zarf-dev/zarf/src/pkg/feature"
 	"github.com/zarf-dev/zarf/src/pkg/logger"
-	"github.com/zarf-dev/zarf/src/pkg/ocitransport"
+	"github.com/zarf-dev/zarf/src/pkg/ocischeme"
 )
 
 var (
@@ -98,7 +98,7 @@ func preRun(cmd *cobra.Command, _ []string) error {
 	}
 	ctx := logger.WithContext(cmd.Context(), l)
 	// One negotiator per CLI invocation; decisions never expire (TTL 0) since Zarf is a short-lived process
-	ctx = ocitransport.WithNegotiator(ctx, ocitransport.New(ocitransport.Options{}))
+	ctx = ocischeme.WithNegotiator(ctx, ocischeme.New(ocischeme.Options{}))
 	cmd.SetContext(ctx)
 
 	// Print enabled features once we have a logger available
