+++ src/pkg/images/push.go
@@ -23,7 +23,7 @@ import (
 	"github.com/zarf-dev/zarf/src/internal/dns"
 	"github.com/zarf-dev/zarf/src/pkg/cluster"
 	"github.com/zarf-dev/zarf/src/pkg/logger"
-	"github.com/zarf-dev/zarf/src/pkg/ocitransport"
+	"github.com/zarf-dev/zarf/src/pkg/ocischeme"
 	"github.com/zarf-dev/zarf/src/pkg/pki"
 	"github.com/zarf-dev/zarf/src/pkg/state"
 	"github.com/zarf-dev/zarf/src/pkg/transform"
@@ -140,7 +140,7 @@ func Push(ctx context.Context, imageList []transform.Image, sourceDirectory stri
 			// mTLS client-certificate transport for a cluster-tunneled registry),
 			// but stripped of any retry wrapper: probing must stay fast, not retry
 			// with backoff on every connection failure.
-			plainHTTP, err = ocitransport.From(ctx).UsePlainHTTP(ctx, registryRef.Host(), ocitransport.ProbeOptions{InsecureSkipTLSVerify: cfg.InsecureSkipTLSVerify, Transport: unwrapRetryTransport(transport)})
+			plainHTTP, err = ocischeme.From(ctx).UsePlainHTTP(ctx, registryRef.Host(), ocischeme.ProbeOptions{InsecureSkipTLSVerify: cfg.InsecureSkipTLSVerify, Transport: unwrapRetryTransport(transport)})
 			if err != nil {
 				return err
 			}
